+++ src/main/java/com/growin/marvel/model/MarvelCharacter.java
@@ -2,11 +2,9 @@ package com.growin.marvel.model;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class MarvelCharacter {
-    private BigInteger id;
+    private int id;
     private String name;
     private String description;
     private Thumbnail thumbnail;
